requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-telegram-bot>=21.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
